+++ include/types.h
@@ -1,5 +1,5 @@
 /* ============================================================================
- * types.h — Basic integer types for freestanding environment
+ * types.h — Basic integer types for freestanding 64-bit environment
  * ============================================================================ */
 
 #pragma once
@@ -7,15 +7,15 @@
 typedef unsigned char       uint8_t;
 typedef unsigned short      uint16_t;
 typedef unsigned int        uint32_t;
-typedef unsigned long long  uint64_t;
+typedef unsigned long       uint64_t;
 
 typedef signed char         int8_t;
 typedef signed short        int16_t;
 typedef signed int          int32_t;
-typedef signed long long    int64_t;
+typedef signed long         int64_t;
 
-typedef uint32_t            size_t;
-typedef int32_t             ssize_t;
-typedef uint32_t            uintptr_t;
+typedef uint64_t            size_t;
+typedef int64_t             ssize_t;
+typedef uint64_t            uintptr_t;
 
 #define NULL ((void *)0)
